[WIP] using common credential fetcher, moving it into OAuth helpers - failing tests
